# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MinuteOps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: programId
          in: query
          schema:
            type: ["integer", "null"]
        - name: schoolId
          in: query
          schema:
            type: ["integer", "null"]
        - name: caseManagerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: riskStatus
          in: query
          schema:
            type: ["string", "null"]
        - name: grade
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentSummary"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}/minute-progress:
    get:
      operationId: getStudentMinuteProgress
      tags: [students]
      summary: Get minute progress for all service requirements of a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Minute progress list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MinuteProgress"
  /students/{id}/sessions:
    get:
      operationId: getStudentSessions
      tags: [students]
      summary: Get recent sessions for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SessionLog"
  /students/{id}/alerts:
    get:
      operationId: getStudentAlerts
      tags: [students]
      summary: Get active alerts for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  # ─── Staff ───────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ─── Schools & Programs ───────────────────────────────────────
...
    StudentSummary:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        externalId:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        placementType:
          type: ["string", "null"]
        status:
          type: string
        schoolId:
          type: ["integer", "null"]
        programId:
          type: ["integer", "null"]
        schoolName:
          type: ["string", "null"]
        programName:
          type: ["string", "null"]
        caseManagerId:
          type: ["integer", "null"]
        caseManagerName:
          type: ["string", "null"]
        riskStatus:
          type: ["string", "null"]
        activeRequirementsCount:
          type: integer
        onTrackCount:
          type: integer
        atRiskCount:
          type: integer
        behindCount:
          type: integer
      required: [id, firstName, lastName, status, activeRequirementsCount, onTrackCount, atRiskCount, behindCount]
    Student:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        externalId:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        placementType:
          type: ["string", "null"]
        status:
          type: string
        schoolId:
          type: ["integer", "null"]
        programId:
          type: ["integer", "null"]
        caseManagerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, firstName, lastName, status, createdAt, updatedAt]
    StudentDetail:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        externalId:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        placementType:
          type: ["string", "null"]
        status:
          type: string
        schoolId:
          type: ["integer", "null"]
        programId:
          type: ["integer", "null"]
        caseManagerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        schoolName:
          type: ["string", "null"]
        programName:
          type: ["string", "null"]
        caseManagerName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        serviceRequirements:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRequirement"
        minuteProgress:
          type: array
          items:
            $ref: "#/components/schemas/MinuteProgress"
        recentSessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionLog"
        activeAlerts:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        assignedStaff:
          type: array
          items:
            $ref: "#/components/schemas/StaffAssignment"
      required: [id, firstName, lastName, status, createdAt, updatedAt, serviceRequirements, minuteProgress, recentSessions, activeAlerts, assignedStaff]
    CreateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        externalId:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        placementType:
          type: ["string", "null"]
        status:
          type: string
        schoolId:
          type: ["integer", "null"]
        programId:
          type: ["integer", "null"]
        caseManagerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
      required: [firstName, lastName, status]
    UpdateStudentBody:
      type: object
      properties:
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        placementType:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        schoolId:
          type: ["integer", "null"]
        programId:
          type: ["integer", "null"]
        caseManagerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]

    # ─── Staff ────────────────────────────────────────────────────
...
    ServiceRequirement:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        serviceTypeId:
          type: integer
        serviceTypeName:
          type: ["string", "null"]
        providerId:
          type: ["integer", "null"]
        providerName:
          type: ["string", "null"]
        deliveryType:
          type: string
        requiredMinutes:
          type: integer
        intervalType:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
      required: [id, studentId, serviceTypeId, deliveryType, requiredMinutes, intervalType, startDate, active, createdAt]
...
    SessionLog:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        serviceRequirementId:
          type: ["integer", "null"]
        serviceTypeId:
          type: ["integer", "null"]
        serviceTypeName:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        sessionDate:
          type: string
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        location:
          type: ["string", "null"]
        deliveryMode:
          type: ["string", "null"]
        status:
          type: string
        missedReasonId:
          type: ["integer", "null"]
        missedReasonLabel:
          type: ["string", "null"]
        isMakeup:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, studentId, sessionDate, durationMinutes, status, isMakeup, createdAt]
...
    MinuteProgress:
      type: object
      properties:
        serviceRequirementId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        serviceTypeId:
          type: integer
        serviceTypeName:
          type: string
        providerId:
          type: ["integer", "null"]
        providerName:
          type: ["string", "null"]
        intervalType:
          type: string
        requiredMinutes:
          type: integer
        deliveredMinutes:
          type: integer
        remainingMinutes:
          type: integer
        percentComplete:
          type: number
        expectedMinutesByNow:
          type: number
        projectedMinutes:
          type: number
        riskStatus:
          type: string
        intervalStart:
          type: string
        intervalEnd:
          type: string
        missedSessionsCount:
          type: integer
        makeupSessionsCount:
          type: integer
      required: [serviceRequirementId, studentId, studentName, serviceTypeId, serviceTypeName, intervalType, requiredMinutes, deliveredMinutes, remainingMinutes, percentComplete, expectedMinutesByNow, projectedMinutes, riskStatus, intervalStart, intervalEnd, missedSessionsCount, makeupSessionsCount]

    # ─── Schedule Blocks ─────────────────────────────────────────
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        severity:
          type: string
        studentId:
          type: ["integer", "null"]
        studentName:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        staffName:
          type: ["string", "null"]
        serviceRequirementId:
          type: ["integer", "null"]
        message:
          type: string
        suggestedAction:
          type: ["string", "null"]
        resolved:
          type: boolean
        resolvedAt:
          type: ["string", "null"]
        resolvedNote:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, type, severity, message, resolved, createdAt]
...
    StaffAssignment:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: ["string", "null"]
        staffRole:
          type: ["string", "null"]
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        assignmentType:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, staffId, studentId, assignmentType, createdAt]
...