# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinuteOps API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  firstName: string;
  lastName: string;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  placementType?: string | null;
  status: string;
  /** @nullable */
  schoolId?: number | null;
  /** @nullable */
  programId?: number | null;
  /** @nullable */
  caseManagerId?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
}